+++ anathema-ssh/Cargo.toml
@@ -15,6 +15,7 @@ anathema-geometry = { workspace = true }
 anathema-widgets = { workspace = true }
 anathema-value-resolver = { workspace = true }
 async-trait = "0.1.88"
+crossterm = { workspace = true }
 futures = "0.3"
 russh = "0.53.0"
 ssh-key = "0.6.7"
