+++ anathema-ssh/src/lib.rs
@@ -1,2 +1,3 @@
 mod sshbackend;
 pub mod sshserver;
+mod terminalhandle;
