+++ examples/ssh.rs
@@ -7,26 +7,29 @@ use anyhow;
 
 #[tokio::main]
 async fn main() -> Result<(), anyhow::Error> {
-    let mut server = AnathemaSSHServer::new(|| {
-        Box::new(move |backend| {
-            let doc = Document::new("@main");
-            let mut builder = Runtime::builder(doc, backend);
-            builder
-                .component(
-                    "main",
-                    "examples/templates/ssh/ssh.aml",
-                    App,
-                    AppState { number: 0.into() },
-                )
-                .unwrap();
-            builder
-                .finish(backend, |runtime, backend| {
-                    println!("RUNTIME RUN...");
-                    runtime.run(backend)
-                })
-                .map_err(|e| anyhow::anyhow!("Runtime error: {}", e))
+    let mut server = AnathemaSSHServer::builder()
+        .runtime_factory(|| {
+            Box::new(move |backend| {
+                let doc = Document::new("@main");
+                let mut builder = Runtime::builder(doc, backend);
+                builder
+                    .component(
+                        "main",
+                        "examples/templates/ssh/ssh.aml",
+                        App,
+                        AppState { number: 0.into() },
+                    )
+                    .unwrap();
+                builder
+                    .finish(backend, |runtime, backend| {
+                        println!("RUNTIME RUN...");
+                        runtime.run(backend)
+                    })
+                    .map_err(|e| anyhow::anyhow!("Runtime error: {}", e))
+            })
         })
-    });
+        .enable_mouse()
+        .build();
 
     println!("Starting SSH server...");
     server.run().await?;
